// src/presentation/Enums/OTFormsEnum.ts
export enum OTFormsEnum {
  folio = 1,
  motivo = 2,
  area_id = 3,
  area = 4,
  estado_id = 5,
  estado = 6,
  validar_parametrizacion_id = 7,
  validar_parametrizacion = 8,
  estado_impresion_id = 9,
  estado_impresion = 10,
  proyecto_codigo = 11,
  proyecto_titulo = 12,
  establecimiento_id = 13,
  establecimiento = 14,

  cliente_rut = 15,
  cliente_nomnbre = 16,
  cliente_tipo = 17,
  cliente_sexo = 18,
  cliente_fecha_nacimiento = 19,
  cliente_direccion = 20,
  cliente_region_id = 21,
  cliente_region = 22,
  cliente_provincia_id = 23,
  cliente_provincia = 24,
  cliente_comuna_id = 25,
  cliente_comuna = 26,
  cliente_telefono = 27,
  cliente_correo = 28,
  oftalmologo_id = 29,
  oftalmologo = 30,

  fecha_atencion = 31,
  fecha_entrega_taller = 32,
  fecha_despacho = 33,
  fecha_entrega_cliente = 34,
  punto_venta_id = 35,
  punto_venta = 36,
  numero_receta = 37,
  fecha_receta = 38,
  tipo_anteojo_id = 39,
  tipo_anteojo = 40,
  a1_od_esf = 41,
  a1_od_cil = 42,
  a1_od_eje = 43,
  a1_od_ad = 44,
  a1_oi_esf = 45,
  a1_oi_cil = 46,
  a1_oi_eje = 47,
  a1_oi_ad = 48,
  a1_dp = 49,
  a1_alt = 50,
  a1_grupo_od = 51,
  a1_grupo_oI = 52,
  a2_od_esf = 53,
  a2_od_cil = 54,
  a2_od_eje = 55,
  a2_oi_esf = 56,
  a2_oi_cil = 57,
  a2_oi_eje = 58,
  a2_dp = 59,
  a2_grupo_od = 60,
  a2_grupo_oi = 61,

  a1_armazon_id = 62,
  a1_armazon = 63,

  a2_armazon_id = 64,
  a2_armazon = 65,

  a3_armazon_id = 66,
  a3_armazon = 67,

  //CRISTALES ANTEOJO 1

  cristal1_marca_id = 68,
  cristal1_marca = 69,
  cristal1_diseno_id = 70,
  cristal1_diseno = 71,
  cristal1_indice_id = 72,
  cristal1_indice = 73,
  cristal1_material_id = 74,
  cristal1_material = 75,
  cristal1_tratamiento_id = 76,
  cristal1_tratamiento = 77,
  cristal1_color_id = 78,
  cristal1_color = 79,
  cristal1_diametro = 80,
  cristal1_od = 81,
  cristal1_oi = 82,

  cristal1_tratamiento_adicional_id = 83,
  cristal1_tratamiento_adicional = 84,

  //CRISTALES ANTEOJO 2

  cristal2_marca_id = 85,
  cristal2_marca = 86,
  cristal2_diseno_id = 87,
  cristal2_diseno = 88,
  cristal2_indice_id = 89,
  cristal2_indice = 90,
  cristal2_material_id = 91,
  cristal2_material = 92,
  cristal2_tratamiento_id = 93,
  cristal2_tratamiento = 94,
  cristal2_color_id = 95,
  cristal2_color = 96,
  cristal2_diametro = 97,
  cristal2_od = 98,
  cristal2_oi = 99,
  cristal2_tratamiento_adicional_id = 100,
  cristal2_tratamiento_adicional = 101,

  //OPTICA
  motivo_garantia_id = 102,
  motivo_garantia = 103,
  folio_asociado = 104,

  resolucion_garantia_id = 105,
  resolucion_garantia = 106,
  worktracking = 107,
  nota_venta = 108,
  numero_reporte_firma = 109,
  numero_reporte_atencion = 110,
  numero_oc = 111,
  numero_guia = 112,
  numero_factura = 113,
  folio_interno_mandante = 114,
  reporte_interno_mandante = 115,
  numero_envio = 116,
  total = 117,
  observaciones = 118,
  nombre_logo = 119,
  imprime_qr = 120,
  imprime_ticket = 121,
  titulo1_ticket = 122,
  titulo2_ticket = 123,
  titulo3_ticket = 124,
  lugar_despacho = 125,
  estado_validacion = 126,
  rbd_establecimiento = 127,
  bodega_procesado = 128,
  usuario_id = 129,

  ot_ubicacion = 130,

  ubicacion_armazon_1 = 131,
  ubicacion_armazon_2 = 132,
  ubicacion_cristal_1_od = 133,
  ubicacion_cristal_1_oi = 134,
  ubicacion_cristal_2_od = 135,
  ubicacion_cristal_2_oi = 136,
  cant_rbd = 137,
  rbd_ubicacion = 138,

  imagen_logo = 139,
}